devops_dj/example1/requirements.txt
Django>=3.0,<=3.0.5
djangorestframework>=3.9.0,<=3.11.0
flake8>=3.6.0,<=3.7.9
psycopg2_binary==2.8.5
django-rest-knox>=4.0,<=4.1.0
gunicorn==20.0.4
xmltodict==0.12.0
Pillow==7.1.2
dnspython3==1.15.0
requests>=2.20
sentry-sdk==0.14.3
django-cors-headers==3.2.1
pytest-cov==2.8.1
codecov==2.0.22
pytest==5.4.2
celery==4.4.2
redis==3.5.2